Support document IDs above signed 32-bit

+++ gomobile/Merge_test.go
@@ -557,7 +557,7 @@ var leftDB = &model.Database{
 		nil,
 		{
 			LocationID:   5,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -727,7 +727,7 @@ var rightDB = &model.Database{
 		},
 		{
 			LocationID:   4,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -907,7 +907,7 @@ var mergedAllLeftDB = &model.Database{
 		},
 		{
 			LocationID:   4,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -1116,7 +1116,7 @@ var mergedAllRightDB = &model.Database{
 		},
 		{
 			LocationID:   4,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -1271,7 +1271,7 @@ var leftNwtDB = &model.Database{
 		nil,
 		{
 			LocationID:   5,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -1347,7 +1347,7 @@ var rightNwtDB = &model.Database{
 		},
 		{
 			LocationID:   4,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -1423,7 +1423,7 @@ var mergedAllLeftNwtDB = &model.Database{
 		},
 		{
 			LocationID:   3,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 			LocationType: 0,
@@ -1504,13 +1504,13 @@ var leftDBNwtWithDifferentDocID = &model.Database{
 		nil,
 		{
 			LocationID:   1,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
 		{
 			LocationID:   2,
-			DocumentID:   sql.NullInt32{123456789, true},
+			DocumentID:   sql.NullInt64{123456789, true},
 			KeySymbol:    sql.NullString{"nwtsty", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
@@ -1583,7 +1583,7 @@ var rightDBNwtWithDifferentDocID = &model.Database{
 		nil,
 		{
 			LocationID:   1,
-			DocumentID:   sql.NullInt32{987654332, true},
+			DocumentID:   sql.NullInt64{987654332, true},
 			KeySymbol:    sql.NullString{"nwt", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
@@ -1663,19 +1663,19 @@ var mergedDBNwtWithDifferentDocID = &model.Database{
 		nil,
 		{
 			LocationID:   1,
-			DocumentID:   sql.NullInt32{1102021811, true},
+			DocumentID:   sql.NullInt64{1102021811, true},
 			KeySymbol:    sql.NullString{"lffi", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
 		{
 			LocationID:   2,
-			DocumentID:   sql.NullInt32{123456789, true},
+			DocumentID:   sql.NullInt64{123456789, true},
 			KeySymbol:    sql.NullString{"nwtsty", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
 		{
 			LocationID:   3,
-			DocumentID:   sql.NullInt32{987654332, true},
+			DocumentID:   sql.NullInt64{987654332, true},
 			KeySymbol:    sql.NullString{"nwt", true},
 			MepsLanguage: sql.NullInt32{Int32: 2, Valid: true},
 		},
